refactor: LineService의 Command, Query를 분리
