+++ src/main/java/nextstep/subway/applicaion/LineQueryService.java
@@ -1,0 +1,33 @@
+package nextstep.subway.applicaion;
+
+import lombok.RequiredArgsConstructor;
+import nextstep.subway.applicaion.dto.LineResponse;
+import nextstep.subway.domain.Line;
+import nextstep.subway.domain.LineRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Service
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+public class LineQueryService {
+
+    private static final String NOT_EXIST_LINE = "존재하지 않는 지하철 노선입니다.";
+
+    private final LineRepository lineRepository;
+
+    public LineResponse findLine(Long id) {
+        Line line = lineRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException(NOT_EXIST_LINE));
+        return LineResponse.from(line);
+    }
+
+    public List<LineResponse> findAllLines() {
+        return lineRepository.findAll().stream()
+                .map(LineResponse::from)
+                .collect(Collectors.toList());
+    }
+}
